Move departuresWithLineName logic
